Pass OS version directly via AppInsights, instead of by Context Property
