+++ src/AccessibilityInsights.Win32/Win32Helper.cs
@@ -194,28 +194,6 @@ namespace Axe.Windows.Win32
         }
 
         /// <summary>
-        /// Get the current Windows version (intended to be used for telemetry). If trying to 
-        /// determine a minimum build for features, please use the IsWindowsXXXOrLater methods
-        /// (or create a new one if you can't find an appropriate one).
-        /// </summary>
-        /// <returns>A string in the format X.Y[.Z] where X.Y is the registry-based CurrentVersion
-        /// and Z is the registry-based CurrentBuild</returns>
-        internal static string GetCurrentWindowsVersionForTelemetry()
-        {
-            string currentVersion = GetCurrentWindowsVersion();
-
-            if (string.IsNullOrWhiteSpace(currentVersion))
-                return null;
-
-            string currentBuild = GetCurrentWindowsBuild();
-
-            if (string.IsNullOrWhiteSpace(currentBuild))
-                return currentVersion;
-
-            return currentVersion + "." + currentBuild;
-        }
-
-        /// <summary>
         /// Get RGB value
         /// </summary>
         /// <param name="r"></param>
